Create toyprogramfactorstinynumbersinstantly.cpp
again, main purpose is to show The Rottura Algorithm #3 which is a "hollow squares" geometric factorization we consider asking if it can be like O(log N) but really direct and O(1) depending how you define time complexity... No iterations, period... And yes, it aught to scale to any size N, and yes it aught to solve PI, and yes it has some chance to, if it scales, lead to compressing the as-yet uncompressible; this program sucks, but meh.  Anyway... WORKS WITH SMALL TEST NUMBERS LIKE 2059 ETC, aught to derive PI to scale of input N, etc.